very-very important update

+++ interfaces/ipc/iface_mdb_coin.py
@@ -26,11 +26,11 @@ class ifaceMDBcoin(ifaceCOINipc,ifaceMDBslave,ifacePAYMENTslave):
 
 	def centsToInternal(self,v):
 		dp=int(10**self.decimal_places)
-		return v/dp/self.scaling_factor
+		return int(v/dp/self.scaling_factor)
 
 	def internalToCents(self,v):
 		dp=int(10**self.decimal_places)
-		return v*dp*self.scaling_factor
+		return int(v*dp*self.scaling_factor)
 
 	def setup(self):
 		response=self.cmd([0x01])
